requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.62.0
tensorboard>=2.6.0
rasterio>=1.2.0
scipy>=1.7.0